saas-store-sync/backend/requirements.txt
django>=4.2,<5.0
djangorestframework>=3.14
djangorestframework-simplejwt>=5.3
django-cors-headers>=4.3
celery>=5.3
redis>=5.0
python-dotenv>=1.0
psycopg2-binary>=2.9
openpyxl>=3.1
cryptography>=41.0
gunicorn>=21.0
requests>=2.31
curl_cffi>=0.7
# Scrapers — Amazon uses Selenium, eBay uses curl_cffi (browser TLS) + BS4
# Optional: pip install pillow && pip install amazoncaptcha --no-deps
selenium>=4.15
webdriver-manager>=4.0
beautifulsoup4>=4.12
lxml>=5.0
